Refactor solution for task BJS2-4510.
