+++ src/main/java/faang/school/godbless/groupingUsersByAge/Main.java
@@ -1,4 +1,4 @@
-package faang.school.godbless;
+package faang.school.godbless.groupingUsersByAge;
 
 import java.util.ArrayList;
 import java.util.List;
